requirements.txt
openai>=1.30.0
streamlit>=1.32.0
pandas>=2.0.0
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dateutil>=2.9.0
Pillow>=10.0.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
pytz>=2023.3
